intf for cfg reloading, jdocs
